refactoring

+++ src/main/java/it/polimi/parkingService/webApplication/utils/BaseService.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface BaseService<T> {
     List<T> findAll();
 
-    T findById(int id);
+    T findById(long id);
 
     T save(T entity);
 
-    void deleteById(int id);
+    void deleteById(long id);
 }
